@@ -114,10 +114,9 @@
     "@rtsao/plugin-proposal-class-properties": "7.0.1-patch.1",
     "@swc/core": "1.3.38",
     "@swc/helpers": "0.4.14",
-    "@testing-library/dom": "8.20.0",
+    "@testing-library/dom": "9.0.1",
     "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
     "@types/angular": "1.8.4",
     "@types/angular-route": "1.7.2",
@@ -146,15 +145,15 @@
     "@types/papaparse": "5.3.7",
     "@types/pluralize": "^0.0.29",
     "@types/prismjs": "1.26.0",
-    "@types/react": "17.0.42",
+    "@types/react": "18.0.28",
     "@types/react-beautiful-dnd": "13.1.3",
-    "@types/react-dom": "17.0.14",
+    "@types/react-dom": "18.0.11",
     "@types/react-grid-layout": "1.3.2",
     "@types/react-highlight-words": "0.16.4",
     "@types/react-redux": "7.1.25",
     "@types/react-router-dom": "5.3.3",
     "@types/react-table": "7.7.14",
-    "@types/react-test-renderer": "17.0.1",
+    "@types/react-test-renderer": "18.0.0",
     "@types/react-transition-group": "4.4.5",
     "@types/react-virtualized-auto-sizer": "1.0.1",
     "@types/react-window": "1.8.5",
@@ -226,7 +225,7 @@
     "react-refresh": "0.14.0",
     "react-select-event": "5.5.1",
     "react-simple-compat": "1.2.3",
-    "react-test-renderer": "17.0.2",
+    "react-test-renderer": "18.2.0",
     "redux-mock-store": "1.5.4",
     "rimraf": "4.4.0",
     "rudder-sdk-js": "2.25.0",
@@ -294,6 +293,7 @@
     "@sentry/browser": "6.19.7",
     "@sentry/types": "6.19.7",
     "@sentry/utils": "6.19.7",
+    "@testing-library/react-hooks": "^8.0.1",
     "@types/react-resizable": "3.0.3",
     "@types/webpack-env": "1.18.0",
     "@visx/event": "3.0.1",
@@ -366,11 +366,11 @@
     "rc-time-picker": "3.7.3",
     "rc-tree": "5.7.2",
     "re-resizable": "6.9.9",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-awesome-query-builder": "5.4.0",
     "react-beautiful-dnd": "13.1.1",
     "react-diff-viewer": "^3.1.1",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
     "react-draggable": "4.4.5",
     "react-dropzone": "^14.2.3",
     "react-enable": "^3.1.0",
@@ -385,7 +385,7 @@
     "react-redux": "7.2.6",
     "react-resizable": "3.0.4",
     "react-reverse-portal": "2.1.1",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "5.3.3",
     "react-select": "5.7.0",
     "react-split-pane": "0.1.92",
     "react-table": "7.8.0",
@@ -437,7 +437,9 @@
     "@storybook/manager-webpack5/webpack": "5.76.0",
     "ngtemplate-loader/loader-utils": "^2.0.0",
     "trim": "0.0.3",
-    "slate-dev-environment@^0.2.2": "patch:slate-dev-environment@npm:0.2.5#.yarn/patches/slate-dev-environment-npm-0.2.5-9aeb7da7b5.patch"
+    "slate-dev-environment@^0.2.2": "patch:slate-dev-environment@npm:0.2.5#.yarn/patches/slate-dev-environment-npm-0.2.5-9aeb7da7b5.patch",
+    "react-split-pane@0.1.92": "patch:react-split-pane@npm:0.1.92#.yarn/patches/react-split-pane-npm-0.1.92-93dbf51dff.patch",
+    "@storybook/addon-docs@6.5.16": "patch:@storybook/addon-docs@npm:6.5.16#.yarn/patches/@storybook-addon-docs-npm-6.5.16-56ecbd77e7.patch"
   },
   "workspaces": {
     "packages": [
