@@ -97,7 +97,7 @@
     "react-inlinesvg": "3.0.2",
     "react-popper": "2.3.0",
     "react-popper-tooltip": "4.4.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "5.3.3",
     "react-select": "5.7.0",
     "react-select-event": "^5.1.0",
     "react-table": "7.8.0",
@@ -134,10 +134,9 @@
     "@storybook/preset-scss": "1.0.3",
     "@storybook/react": "6.5.16",
     "@storybook/theming": "6.5.16",
-    "@testing-library/dom": "8.20.0",
+    "@testing-library/dom": "9.0.1",
     "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
     "@types/common-tags": "^1.8.0",
     "@types/d3": "7.4.0",
@@ -149,15 +148,15 @@
     "@types/mock-raf": "1.0.3",
     "@types/node": "18.14.6",
     "@types/prismjs": "1.26.0",
-    "@types/react": "17.0.42",
+    "@types/react": "18.0.28",
     "@types/react-beautiful-dnd": "13.1.3",
     "@types/react-calendar": "3.9.0",
     "@types/react-color": "3.0.6",
-    "@types/react-dom": "17.0.14",
+    "@types/react-dom": "18.0.11",
     "@types/react-highlight-words": "0.16.4",
     "@types/react-router-dom": "5.3.3",
     "@types/react-table": "7.7.14",
-    "@types/react-test-renderer": "17.0.1",
+    "@types/react-test-renderer": "18.0.0",
     "@types/react-transition-group": "4.4.5",
     "@types/react-window": "1.8.5",
     "@types/slate": "0.47.11",
@@ -173,9 +172,9 @@
     "expose-loader": "4.0.0",
     "mock-raf": "1.0.1",
     "process": "^0.11.10",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-test-renderer": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-test-renderer": "18.2.0",
     "rimraf": "4.4.0",
     "rollup": "2.79.1",
     "rollup-plugin-dts": "^5.0.0",
@@ -190,7 +189,7 @@
     "webpack": "5.76.0"
   },
   "peerDependencies": {
-    "react": "^16.8.0 || ^17.0.0",
-    "react-dom": "^16.8.0 || ^17.0.0"
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
   }
 }
