@@ -57,7 +57,7 @@ describe('ErrorBoundary', () => {
     expect((faro.api.pushError as jest.Mock).mock.calls[0][0]).toBe(problem);
   });
 
-  it('should recover when when recover props change', async () => {
+  it('should rerender when recover props change', async () => {
     const problem = new Error('things went terribly wrong');
     let renderCount = 0;
 
@@ -75,6 +75,8 @@ describe('ErrorBoundary', () => {
     );
 
     await screen.findByText(problem.message);
+    expect(renderCount).toBeGreaterThan(0);
+    const oldRenderCount = renderCount;
 
     rerender(
       <ErrorBoundary dependencies={[1, 3]}>
@@ -89,6 +91,6 @@ describe('ErrorBoundary', () => {
       </ErrorBoundary>
     );
 
-    expect(renderCount).toBe(2);
+    expect(renderCount).toBeGreaterThan(oldRenderCount);
   });
 });
