@@ -1,6 +1,4 @@
-import { fireEvent } from '@testing-library/dom';
-import { render, screen } from '@testing-library/react';
-import { act, renderHook } from '@testing-library/react-hooks';
+import { act, fireEvent, render, renderHook, screen } from '@testing-library/react';
 import React, { createRef, KeyboardEvent, RefObject } from 'react';
 
 import { useMenuFocus } from './hooks';
