@@ -357,7 +357,11 @@ export const LogRowContextGroup = ({
                         padding: 5px 0;
                       `}
                     >
-                      {typeof item === 'string' && textUtil.hasAnsiCodes(item) ? <LogMessageAnsi value={item} /> : item}
+                      {typeof item === 'string' && textUtil.hasAnsiCodes(item) ? (
+                        <LogMessageAnsi value={item} />
+                      ) : (
+                        String(item)
+                      )}
                     </div>
                   );
                 }}
