@@ -71,7 +71,7 @@ const QueryEditor = ({
         <>
           <Space v={2} />
           <Alert severity="error" title="An error occurred while requesting metadata from Azure Monitor">
-            {errorMessage}
+            {errorMessage instanceof Error ? errorMessage.message : errorMessage}
           </Alert>
         </>
       )}
