@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook, waitFor } from '@testing-library/react';
 
 import { config } from '@grafana/runtime';
 
@@ -12,10 +12,6 @@ import {
 } from './__mocks__/CloudWatchDataSource';
 import { setupMockedResourcesAPI } from './__mocks__/ResourcesAPI';
 import { useAccountOptions, useDimensionKeys, useIsMonitoringAccount, useMetrics } from './hooks';
-
-const WAIT_OPTIONS = {
-  timeout: 1000,
-};
 
 const originalFeatureToggleValue = config.featureToggles.cloudWatchCrossAccountQuerying;
 
@@ -32,10 +28,11 @@ describe('hooks', () => {
       const isMonitoringAccountMock = jest.fn().mockResolvedValue(true);
       api.isMonitoringAccount = isMonitoringAccountMock;
 
-      const { waitForNextUpdate } = renderHook(() => useIsMonitoringAccount(api, `$${regionVariable.name}`));
-      await waitForNextUpdate(WAIT_OPTIONS);
-      expect(isMonitoringAccountMock).toHaveBeenCalledTimes(1);
-      expect(isMonitoringAccountMock).toHaveBeenCalledWith(regionVariable.current.value);
+      renderHook(() => useIsMonitoringAccount(api, `$${regionVariable.name}`));
+      await waitFor(() => {
+        expect(isMonitoringAccountMock).toHaveBeenCalledTimes(1);
+        expect(isMonitoringAccountMock).toHaveBeenCalledWith(regionVariable.current.value);
+      });
     });
   });
   describe('useMetricNames', () => {
@@ -46,19 +43,20 @@ describe('hooks', () => {
       const getMetricsMock = jest.fn().mockResolvedValue([]);
       datasource.resources.getMetrics = getMetricsMock;
 
-      const { waitForNextUpdate } = renderHook(() =>
+      renderHook(() =>
         useMetrics(datasource, {
           namespace: `$${namespaceVariable.name}`,
           region: `$${regionVariable.name}`,
           accountId: `$${accountIdVariable.name}`,
         })
       );
-      await waitForNextUpdate(WAIT_OPTIONS);
-      expect(getMetricsMock).toHaveBeenCalledTimes(1);
-      expect(getMetricsMock).toHaveBeenCalledWith({
-        region: regionVariable.current.value,
-        namespace: namespaceVariable.current.value,
-        accountId: accountIdVariable.current.value,
+      await waitFor(() => {
+        expect(getMetricsMock).toHaveBeenCalledTimes(1);
+        expect(getMetricsMock).toHaveBeenCalledWith({
+          region: regionVariable.current.value,
+          namespace: namespaceVariable.current.value,
+          accountId: accountIdVariable.current.value,
+        });
       });
     });
   });
@@ -72,7 +70,7 @@ describe('hooks', () => {
       const getDimensionKeysMock = jest.fn().mockResolvedValue([]);
       datasource.resources.getDimensionKeys = getDimensionKeysMock;
 
-      const { waitForNextUpdate } = renderHook(() =>
+      renderHook(() =>
         useDimensionKeys(datasource, {
           namespace: `$${namespaceVariable.name}`,
           metricName: `$${metricVariable.name}`,
@@ -83,16 +81,17 @@ describe('hooks', () => {
           },
         })
       );
-      await waitForNextUpdate(WAIT_OPTIONS);
-      expect(getDimensionKeysMock).toHaveBeenCalledTimes(1);
-      expect(getDimensionKeysMock).toHaveBeenCalledWith({
-        region: regionVariable.current.value,
-        namespace: namespaceVariable.current.value,
-        metricName: metricVariable.current.value,
-        accountId: accountIdVariable.current.value,
-        dimensionFilters: {
-          environment: [dimensionVariable.current.value],
-        },
+      await waitFor(() => {
+        expect(getDimensionKeysMock).toHaveBeenCalledTimes(1);
+        expect(getDimensionKeysMock).toHaveBeenCalledWith({
+          region: regionVariable.current.value,
+          namespace: namespaceVariable.current.value,
+          metricName: metricVariable.current.value,
+          accountId: accountIdVariable.current.value,
+          dimensionFilters: {
+            environment: [dimensionVariable.current.value],
+          },
+        });
       });
     });
   });
@@ -105,9 +104,10 @@ describe('hooks', () => {
       });
       const getAccountsMock = jest.fn().mockResolvedValue([{ id: '123', label: 'accountLabel' }]);
       api.getAccounts = getAccountsMock;
-      const { waitForNextUpdate } = renderHook(() => useAccountOptions(api, `$${regionVariable.name}`));
-      await waitForNextUpdate(WAIT_OPTIONS);
-      expect(getAccountsMock).toHaveBeenCalledTimes(0);
+      renderHook(() => useAccountOptions(api, `$${regionVariable.name}`));
+      await waitFor(() => {
+        expect(getAccountsMock).toHaveBeenCalledTimes(0);
+      });
     });
 
     it('interpolates region variables before calling the api', async () => {
@@ -117,10 +117,11 @@ describe('hooks', () => {
       });
       const getAccountsMock = jest.fn().mockResolvedValue([{ id: '123', label: 'accountLabel' }]);
       api.getAccounts = getAccountsMock;
-      const { waitForNextUpdate } = renderHook(() => useAccountOptions(api, `$${regionVariable.name}`));
-      await waitForNextUpdate(WAIT_OPTIONS);
-      expect(getAccountsMock).toHaveBeenCalledTimes(1);
-      expect(getAccountsMock).toHaveBeenCalledWith({ region: regionVariable.current.value });
+      renderHook(() => useAccountOptions(api, `$${regionVariable.name}`));
+      await waitFor(() => {
+        expect(getAccountsMock).toHaveBeenCalledTimes(1);
+        expect(getAccountsMock).toHaveBeenCalledWith({ region: regionVariable.current.value });
+      });
     });
 
     it('returns properly formatted account options, and template variables', async () => {
@@ -130,12 +131,13 @@ describe('hooks', () => {
       });
       const getAccountsMock = jest.fn().mockResolvedValue([{ id: '123', label: 'accountLabel' }]);
       api.getAccounts = getAccountsMock;
-      const { waitForNextUpdate, result } = renderHook(() => useAccountOptions(api, `$${regionVariable.name}`));
-      await waitForNextUpdate(WAIT_OPTIONS);
-      expect(result.current.value).toEqual([
-        { label: 'accountLabel', description: '123', value: '123' },
-        { label: 'Template Variables', options: [{ label: '$region', value: '$region' }] },
-      ]);
+      const { result } = renderHook(() => useAccountOptions(api, `$${regionVariable.name}`));
+      await waitFor(() => {
+        expect(result.current.value).toEqual([
+          { label: 'accountLabel', description: '123', value: '123' },
+          { label: 'Template Variables', options: [{ label: '$region', value: '$region' }] },
+        ]);
+      });
     });
   });
 });
