@@ -1,5 +1,4 @@
-import { fireEvent, screen } from '@testing-library/dom';
-import { render } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import React, { useState } from 'react';
 
 import { CoreApp, MutableDataFrame } from '@grafana/data';
